requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-nmap>=0.7.1
dnspython>=2.2.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
tqdm>=4.62.0
colorama>=0.4.4
google-generativeai>=0.3.0
selenium>=4.0.0
urllib3>=1.26.0
lxml>=4.9.0